requirements.txt
flask==2.3.2
flask-socketio==5.3.4
openai==1.3.5
google-genai==0.1.0
python-dotenv==1.0.0
reportlab==4.0.4
werkzeug==2.3.6